Create hcpcapacityreporting_xml.py
This is the exact same script as the previous hcp reporting. But I have used the inbuilt xml module to fetch information. I happen to notice that the lxml module causes a lot of issues in some platforms. So to make it easy I developed a fall back option of the same script.